finished fixing the intent builders, optimized api requests
