+++ src/backend/reponse_builder/full_ruletext_response_builder.py
@@ -1,0 +1,15 @@
+from ..Spell_class import Spell_class
+
+
+def build_full_ruletext(spell_name_in):
+    spell = Spell_class(spell_name_in)
+    response = ''
+
+    if spell.name == 'empty':
+        return 'empty'
+    else:
+        if spell.desc != 'empty':
+            response = spell.desc
+        else:
+            response = 'No description found'
+    return response
