+++ src/backend/spell_api_wrapper.py
@@ -19,15 +19,19 @@ class Spell_api_wrapper():
             self._spell_name = spell_name_in.replace(' ', '-')
         else:
             self._spell_name = 'empty'
-        if self.api_reachable() == False:
+        self._api_reachable = self.api_reachable()
+        if self._api_reachable:
+            self.response = self.api_request()
+        else:
             raise 'api not availlable on path'
+        
 
     """
     returns true if api is reachable
     """
     def api_reachable(self):
         try:
-            response = requests.get(self._api_path, timeout=5)
+            response = requests.get(self._api_path, timeout=3)
             response.raise_for_status()
             return True
         except requests.exceptions.HTTPError as errh:
@@ -66,12 +70,11 @@ class Spell_api_wrapper():
     - if required, an index-/ range can be passed
     """
     def get_detail(self, key):
-        if self._spell_name == 'empty':
+        if (self._spell_name == 'empty') or (not self._api_reachable):
             parsed_response = 'empty'
         else:
-            response = self.api_request()
-            if response is not None:
-                response_json = response.json()
+            if self.response is not None:
+                response_json = self.response.json()
                 parsed_response = functools.reduce(dict.get, key, response_json)
                 if parsed_response is None:
                     parsed_response = 'empty'
@@ -79,24 +82,3 @@ class Spell_api_wrapper():
                 parsed_response = 'empty'
 
         return parsed_response
-
-
-'''
-    def get_detail(self, key, index_start = -1, index_stop = -1):
-        if self._spell_name == 'empty':
-            parsed_response = 'empty'
-        else:
-            response = self.api_request()
-            if response is not None:
-                response_json = response.json()
-                parsed_response = functools.reduce(dict.get, key, response_json)
-                if index_start is not -1:
-                    if index_stop is not -1:
-                        parsed_response = list(parsed_response.values())[index_start:index_stop]  
-                    else:  
-                        parsed_response = list(parsed_response.values())[index_start]
-            else:
-                parsed_response = 'empty'
-
-        return parsed_response
-'''
