+++ src/intents/read_all_details.py
@@ -1,4 +1,4 @@
-from ..backend.response_builder import build_response_all_details
+from ..backend.reponse_builder.all_details_response_builder import build_response_all_details
 
 def _read_all_details(self, message):
     spell_name_input = message.data.get('spellname')
